Contents of __init__ moved to config.py
