+++ tests/test_io.py
@@ -3,7 +3,7 @@ import xarray as xr
 import numpy as np
 import json
 
-from agage_archive import Paths, open_data_file, data_file_path, data_file_list
+from agage_archive.config import Paths, open_data_file, data_file_path, data_file_list
 from agage_archive.convert import scale_convert
 from agage_archive.io import read_ale_gage, combine_datasets, read_nc_path, read_nc, \
     read_baseline, combine_baseline, output_dataset
